Move `BuilderAnnotationProcessor` and related classes to `processor` package.
