+++ processor/src/main/java/io/bothy/tog/processor/ConstructorBuilderTarget.java
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License along with Foobar. If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package io.bothy.tog;
+package io.bothy.tog.processor;
 
 import java.util.List;
 import javax.lang.model.element.Element;
