+++ processor/src/test/resources/class/constructor/expected/MenuBuilder.java
@@ -4,7 +4,7 @@ import java.lang.String;
 import java.util.List;
 import javax.annotation.processing.Generated;
 
-@Generated("io.bothy.tog.BuilderAnnotationProcessor")
+@Generated("io.bothy.tog.processor.BuilderAnnotationProcessor")
 public interface MenuBuilder {
 
     static MenuBuilder builder() {
