+++ processor/src/test/resources/record/multiField/expected/PersonBuilder.java
@@ -4,7 +4,7 @@ import java.lang.String;
 import java.util.List;
 import javax.annotation.processing.Generated;
 
-@Generated("io.bothy.tog.BuilderAnnotationProcessor")
+@Generated("io.bothy.tog.processor.BuilderAnnotationProcessor")
 public interface PersonBuilder {
 
     static PersonBuilder builder() {
